tried out codeplay compiler. Works great!

+++ src/sycl/XSbench_header.h
@@ -65,17 +65,17 @@ typedef struct{
 	double * unionized_energy_array;    // Length = length_unionized_energy_array
 	int * index_grid;                   // Length = length_index_grid
 	NuclideGridPoint * nuclide_grid;    // Length = length_nuclide_grid
-	int length_num_nucs;
-	int length_concs;
-	int length_mats;
-	int length_unionized_energy_array;
-	int length_index_grid;
-	int length_nuclide_grid;
+	long length_num_nucs;
+	long length_concs;
+	long length_mats;
+	long length_unionized_energy_array;
+	long length_index_grid;
+	long length_nuclide_grid;
 	int max_num_nucs;
 	double * p_energy_samples;
-	int length_p_energy_samples;
+	long length_p_energy_samples;
 	int * mat_samples;
-	int length_mat_samples;
+	long length_mat_samples;
 } SimulationData;
 
 // io.c
